pyproject.toml
[project]
name = "tfthud"
version = "0.1.0"
description = "TFT HUD - Asistente de Draft en Tiempo Real"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis>=5.2.0",
    "structlog>=24.4.0",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "apscheduler>=3.11.0",
    "slowapi>=0.1.9",
    "python-dotenv>=1.0.0",
    "tqdm>=4.67.0",
    "mss>=9.0.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[dependency-groups]
dev = [
    "black>=25.0.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest-watch>=4.2.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["I", "E", "F", "W", "N", "UP", "B", "C4", "ANN"]
ignore = ["ANN101", "ANN401"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "scraper*", "shared*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["backend", "scraper", "shared"]
omit = ["tests/*", "**/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]